Fixed warning
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1778 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/hash_map.cpp
@@ -1288,7 +1288,7 @@ void HashMap<KeyType, ValueType>::increaseSize (const int powerOffset)
 				dataIter!=data[i].end();
 				)
 			{
-				if((dataIter->hash & mask)!=i)
+				if((dataIter->hash & mask)!=(unsigned int)i)
 				{
 					map[dataIter->hash & mask]=offset;
 					data[dataIter->hash & mask].push_front(*dataIter);
